Parse etags using a 'straight' XML parse operation
Because Feedparser ignores even namespaced attributes it doesn't
know about, e.g. <content … zapi:etag="1236352"> … </content>
That value has to be extracted using a standard XML parser.
It, as well as other metadata (key, group ID if it exists) are added to
the item dicts in the standard_items() function.