Adding the 'updated' time to item dicts
At some point, I'll implement HTTP 304 support for the item()
API call. In order for that to work, the header has to be sent in
the format the server expects. Item dicts now contain this info
in the 'updated' key, in the correct format, and the key is removed
before dicts are sent back during Write API calls