This is a partial reversion of a fix for #72.
The original fix (b338e44) removed all control characters. However,
this introduced an inconsistency between the way in which Zotero[.org]
represents and exports data, and how Pyzotero presents the data. This
commit removes that inconsistency.

The fix involves the addition of python-bibtexparser, which is now
used to process bibtex entries.