Add linkMode parameter to item template function
Requesting item_template for attachment threw an error from the REST
API indicating that it needs a linkMode parameter in that case. I
added this feature to the item_template function and tested for a use
case for adding links to article identifiers from the xDD library.

I may not have handled the error condition in the best way possible as I didn't fully track down functionality of the error_handler.